Extract test reporter to separate file
