+++ test/docs.test.ts
@@ -1,0 +1,4 @@
+import path from "path";
+import { runTests } from "./runTests";
+
+runTests(path.join(__dirname, "../", "docs"), ".ats");
